+++ src/albert/collections/targets.py
@@ -82,8 +82,8 @@ class TargetCollection(BaseCollection):
                 target=Target(
                     name="Viscosity spec",
                     type=TargetType.PERFORMANCE,
-                    data_template_id="DAT1",
-                    data_column_id="DAC1",
+                    data_template_id="DAT9999999",
+                    data_column_id="DAC9999999",
                     target_value=Criterion(operator=ComparisonOperator.GTE, value=90),
                     is_required=True,
                 )